requirements.txt
# === Core Application Libraries ===
streamlit
streamlit-webrtc
fer
streamlit-audiorecorder
simpleaudio
soundfile
pydub
keyboard
python-dotenv
cryptography

# === Manually Resolved Dependencies (Crucial for Stability) ===
moviepy==1.0.3
tensorflow
opencv-contrib-python

# === API SDKs ===
groq
deepgram-sdk==2.12.0
edge-tts

# === Speech Emotion Recognition ===
transformers
torch
librosa
sentencepiece
onnx
onnxruntime